Extract sync task out of SyncJobService
