+++ app/src/main/java/com/chillibits/pmapp/tool/BootCompletedReceiver.kt
@@ -9,7 +9,7 @@ import android.app.PendingIntent
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import com.chillibits.pmapp.service.SyncJobService
+import com.chillibits.pmapp.service.SyncService
 
 class BootCompletedReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context, intent: Intent) {
@@ -21,7 +21,7 @@ class BootCompletedReceiver : BroadcastReceiver() {
             // Setup AlarmManager
             val am = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
 
-            val startServiceIntent = Intent(context, SyncJobService::class.java)
+            val startServiceIntent = Intent(context, SyncService::class.java)
             startServiceIntent.putExtra("FromBackground", true)
             val startServicePendingIntent = PendingIntent.getService(context, Constants.REQ_ALARM_MANAGER_BACKGROUND_SYNC, startServiceIntent, 0)
             am.setRepeating(AlarmManager.RTC_WAKEUP, System.currentTimeMillis(), backgroundSyncFrequency.toLong(), startServicePendingIntent)
