+++ app/src/main/res/values-bg/strings.xml
@@ -9,7 +9,7 @@
  Release: Working copy
  Locale: bg, Bulgarian
  Exported by: ChilliBits
- Exported at: Tue, 04 Feb 2020 17:51:33 +0000
+ Exported at: Tue, 04 Feb 2020 10:01:36 +0000
 -->
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
     <!-- Common Strings -->
