+++ app/src/main/res/values-de/strings.xml
@@ -9,7 +9,7 @@
  Release: Working copy
  Locale: de, German
  Exported by: ChilliBits
- Exported at: Tue, 04 Feb 2020 18:51:34 +0100
+ Exported at: Tue, 04 Feb 2020 11:01:39 +0100
 -->
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
     <!-- Common Strings -->
