+++ app/src/main/res/values-hu/strings.xml
@@ -9,7 +9,7 @@
  Release: Working copy
  Locale: hu, Hungarian
  Exported by: ChilliBits
- Exported at: Tue, 04 Feb 2020 15:45:06 +0100
+ Exported at: Tue, 04 Feb 2020 11:01:39 +0100
 -->
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
     <!-- Common Strings -->
