+++ app/src/main/res/values-nl/strings.xml
@@ -9,7 +9,7 @@
  Release: Working copy
  Locale: nl, Dutch
  Exported by: ChilliBits
- Exported at: Tue, 04 Feb 2020 14:45:04 +0000
+ Exported at: Tue, 04 Feb 2020 20:46:13 +0000
 -->
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
     <!-- Common Strings -->
