+++ app/src/main/res/values-ro/strings.xml
@@ -9,7 +9,7 @@
  Release: Working copy
  Locale: ro, Romanian
  Exported by: ChilliBits
- Exported at: Tue, 04 Feb 2020 14:45:05 +0000
+ Exported at: Tue, 04 Feb 2020 10:01:38 +0000
 -->
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
     <!-- Common Strings -->
